# swagger.yaml
openapi: 3.0.0
info:
  title: API Toko Online (Point of Sale)
  version: 1.0.0
  description: Dokumentasi API lengkap untuk proyek Toko Online, dibuat sesuai Jobsheet.
servers:
  - url: http://localhost:3000
    description: Server Pengembangan

# 1. Definisi Tag (Grup Endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Masukkan "Bearer [token]" (tanpa kurung). Token didapat dari /auth/login.'
  
  # 3. Definisi Skema (Bentuk Data)
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        fullname:
          type: string
          example: "Administrator"
        role:
          type: string
          example: "admin"
        is_active:
          type: boolean
          example: true
...
paths:
...
  /register:
    post:
      tags: [User Management]
      summary: Registrasi user baru (Hanya Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "kasir02"
                password:
                  type: string
                  example: "passwordkasir"
                fullname:
                  type: string
                  example: "Budi"
                role:
                  type: string
                  example: "kasir"
      responses:
        '201':
          description: Pengguna baru berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Hanya Admin yang bisa
        '409':
          description: Username sudah digunakan
  /profile:
    get:
      tags: [User Management]
      summary: Melihat profil sendiri (Admin/Kasir)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Data profil pengguna
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Token tidak tersedia atau kedaluwarsa
  /deactivate-user:
    put:
      tags: [User Management]
      summary: Menonaktifkan akun user (Hanya Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 3
      responses:
        '200':
          description: Pengguna berhasil dinonaktifkan
        '400':
          description: Tidak dapat menonaktifkan akun sendiri
        '404':
          description: Pengguna tidak ditemukan

  # --- CATEGORIES (ADMIN) ---
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Mengambil semua transaksi (Admin/Kasir)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Daftar semua transaksi
    post:
      tags: [Transactions]
      summary: Membuat transaksi baru (Admin/Kasir)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method:
                  type: string
                  example: "cash"
                customer_name:
                  type: string
                  example: "Bambang"
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product_id:
                        type: integer
                        example: 1
                      quantity:
                        type: integer
                        example: 2
      responses:
        '201':
          description: Transaksi berhasil dibuat
        '400':
          description: Stok tidak mencukupi atau produk tidak ditemukan